+++ 01-Python/04-Solutions/04-Operators/Q01.py
@@ -1,0 +1,40 @@
+"""
+========================================================
+Module   : 01-Python
+Chapter  : 04-Operators
+Question : Q01
+========================================================
+
+Question:
+
+Create two variables:
+
+a = 20
+b = 5
+
+Perform the following operations:
+
+- Addition
+- Subtraction
+- Multiplication
+- Division
+
+Print all results.
+
+--------------------------------------------------------
+Solution
+--------------------------------------------------------
+"""
+
+
+
+# Write your solution below
+
+a = 20
+b = 5
+
+print(f"Addition       : {a+b}")
+print(f"Subtraction    : {a-b}")
+print(f"Multiplication : {a*b}")
+print(f"Division       : {a/b}")
+