+++ 01-Python/04-Solutions/04-Operators/Q09.py
@@ -1,0 +1,32 @@
+"""
+========================================================
+Module   : 01-Python
+Chapter  : 04-Operators
+Question : Q09
+========================================================
+
+Question:
+
+Create a variable called `salary`.
+
+Increase it by ₹5,000 using an assignment operator.
+
+
+--------------------------------------------------------
+Solution
+--------------------------------------------------------
+"""
+
+
+
+# Write your solution below
+
+salary = 20000
+
+salary+=5000
+
+print(salary)
+
+
+
+