+++ 01-Python/04-Solutions/04-Operators/Q15.py
@@ -1,0 +1,28 @@
+"""
+========================================================
+Module   : 01-Python
+Chapter  : 04-Operators
+Question : Q15
+========================================================
+
+Question:
+
+Predict the output.
+
+print(True and False)
+
+
+--------------------------------------------------------
+Solution
+--------------------------------------------------------
+"""
+
+
+
+# Write your solution below
+
+print(True and False)
+
+# false
+
+