+++ 01-Python/04-Solutions/04-Operators/Q16.py
@@ -1,0 +1,31 @@
+"""
+========================================================
+Module   : 01-Python
+Chapter  : 04-Operators
+Question : Q16
+========================================================
+
+Question:
+
+Predict the output.
+
+print(True or False)
+
+
+
+--------------------------------------------------------
+Solution
+--------------------------------------------------------
+"""
+
+
+
+# Write your solution below
+
+print(True or False)
+
+# True
+
+
+
+