+++ 01-Python/04-Solutions/04-Operators/Q19.py
@@ -1,0 +1,32 @@
+"""
+========================================================
+Module   : 01-Python
+Chapter  : 04-Operators
+Question : Q19
+========================================================
+
+Question:
+
+Predict the output.
+
+numbers = [10, 20, 30]
+
+print(50 not in numbers)
+
+
+--------------------------------------------------------
+Solution
+--------------------------------------------------------
+"""
+
+
+
+# Write your solution below
+
+
+numbers = [10, 20, 30]
+
+print(50 not in numbers)
+
+# True
+