+++ 01-Python/04-Solutions/04-Operators/Q20.py
@@ -1,0 +1,39 @@
+"""
+========================================================
+Module   : 01-Python
+Chapter  : 04-Operators
+Question : Q20
+========================================================
+
+Question:
+
+Create two lists having the same values.
+
+Print the result of:
+
+==
+
+and
+
+is
+
+Explain the difference.
+
+
+
+
+--------------------------------------------------------
+Solution
+--------------------------------------------------------
+"""
+
+
+
+# Write your solution below
+
+number_1 = [10, 20, 30]
+number_2 = [10, 20, 30]
+
+print(number_1==number_2)
+print(number_1 is number_2)
+