+++ 01-Python/04-Solutions/04-Operators/Q29.py
@@ -1,0 +1,41 @@
+"""
+========================================================
+Module   : 01-Python
+Chapter  : 04-Operators
+Question : Q29
+========================================================
+
+Question:
+
+An employee receives:
+
+- Basic Salary
+- HRA
+- Bonus
+
+A tax amount is deducted.
+
+Calculate the Net Salary.
+
+
+
+--------------------------------------------------------
+Solution
+--------------------------------------------------------
+"""
+
+
+
+# Write your solution below
+
+
+basic_salary = 34000
+hra = 20000
+bonus = 9400
+
+tax = 8300
+
+print(f"In-hand salary = {basic_salary+hra+bonus-tax}")
+
+
+