+++ 01-Python/04-Solutions/04-Operators/Q30.py
@@ -1,0 +1,44 @@
+"""
+========================================================
+Module   : 01-Python
+Chapter  : 04-Operators
+Question : Q30
+========================================================
+
+Question:
+
+A student has marks in five subjects.
+
+Calculate:
+
+- Total Marks
+- Percentage
+
+Determine whether the student has scored **more than 60%** using a comparison operator.
+
+Print the result.
+
+
+
+--------------------------------------------------------
+Solution
+--------------------------------------------------------
+"""
+
+
+
+# Write your solution below
+
+
+maths = 94
+physics = 84
+chemistry = 95
+english = 87
+biology = 91
+
+
+total_marks = maths+physics+chemistry+english+biology
+
+print(f"Total Marks: {total_marks}")
+print(f"Percentage: {total_marks/5}")
+